backend/requirements.txt
# FastAPI ve Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# OCR Libraries
google-cloud-documentai==2.24.0
boto3==1.34.28  # AWS Textract
openai==1.10.0
aiohttp==3.9.1  # PaddleOCR mikroservis client

# Image Processing
Pillow==10.2.0
opencv-python==4.6.0.66

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tiktoken==0.5.2

# CORS
python-jose[cryptography]==3.3.0
